Safari-611.1.10/LayoutTests/fast/events/window-event-recursive-expected.txt

Tests that window.event is set on the right global when an event is fired.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS otherWindow.event is undefined.
PASS window.event is undefined.
Dispatching error event
PASS otherWindow.event is undefined.
PASS passedEvent is window.event
Dispatching recursive error event
PASS otherWindow.event is undefined.
PASS passedEvent is window.event
PASS otherWindow.event is undefined.
PASS passedEvent is window.event
PASS otherWindow.event is undefined.
PASS window.event is undefined.
PASS successfullyParsed is true

TEST COMPLETE
